refactor: extract kolu-git package from server
Move git operations (worktree CRUD, diff review, repo resolution,
HEAD watching) into packages/integrations/git/ following the existing
integration package pattern. Replace thrown exceptions with sum-type
GitResult<T> returns and inject Logger via optional parameter.

Server keeps a thin provider adapter in meta/git.ts that bridges
kolu-git results into the metadata event system.
