+++ packages/integrations/git/src/safe-path.ts
@@ -1,12 +1,13 @@
 /**
- * Secure path helpers for RPC handlers that receive file paths from the
- * client. A child path arriving over RPC is untrusted — a crafted
- * `../../etc/passwd` would be read by `fs.readFile` or handed to
- * subprocess invocations unless we normalize and reject it up front.
+ * Secure path helpers for operations that receive file paths from
+ * untrusted sources. A child path arriving over RPC is untrusted —
+ * a crafted `../../etc/passwd` would escape the repo root unless
+ * we normalize and reject it up front.
  */
 
 import path from "node:path";
-import { log } from "./log.ts";
+import type { Logger } from "kolu-integration-common";
+import { type GitResult, ok, err } from "./errors.ts";
 
 /**
  * Normalize a caller-supplied `child` path against a trusted `root` and
@@ -15,7 +16,7 @@ import { log } from "./log.ts";
  * every downstream subprocess / tool call — so no code path is reading
  * the raw untrusted string).
  *
- * Uses the `path.relative` idiom (`..` prefix ⇒ outside) rather than a
+ * Uses the `path.relative` idiom (`..` prefix => outside) rather than a
  * `startsWith(rootAbs + path.sep)` prefix check; same guarantee, no
  * trailing-slash gotcha, and we get `rel` out of the computation for
  * free.
@@ -23,13 +24,14 @@ import { log } from "./log.ts";
 export function resolveUnder(
   root: string,
   child: string,
-): { abs: string; rel: string } {
+  log?: Logger,
+): GitResult<{ abs: string; rel: string }> {
   const rootAbs = path.resolve(root);
   const abs = path.resolve(rootAbs, child);
   const rel = path.relative(rootAbs, abs);
   if (rel.startsWith("..") || path.isAbsolute(rel)) {
-    log.error({ root, child }, "safe-path: child escapes root");
-    throw new Error(`path escapes root: ${child}`);
+    log?.error({ root, child }, "safe-path: child escapes root");
+    return err({ code: "PATH_ESCAPES_ROOT", root, child });
   }
-  return { abs, rel };
+  return ok({ abs, rel });
 }
