+++ packages/server/package.json
@@ -26,12 +26,12 @@
     "hono-pino": "^0.10.3",
     "kolu-claude-code": "workspace:*",
     "kolu-common": "workspace:*",
+    "kolu-git": "workspace:*",
     "kolu-opencode": "workspace:*",
     "node-pty": "^1.0.0",
     "pino": "^10.3.1",
     "pino-pretty": "^13.1.3",
     "selfsigned": "^5.5.0",
-    "simple-git": "^3.33.0",
     "string-argv": "^0.3.2",
     "ts-pattern": "^5.9.0",
     "ws": "^8.19.0",
