+++ packages/server/src/meta/git.ts
@@ -10,154 +10,18 @@
  * 3. CWD event in a non-git dir where .git now exists → detects `git init`
  */
 
-import path from "node:path";
-import fs from "node:fs";
-import { execSync } from "node:child_process";
-import { simpleGit } from "simple-git";
 import type { GitInfo } from "kolu-common";
+import {
+  resolveGitInfo,
+  watchGitHead,
+  gitInfoEqual,
+  hasGitDir,
+} from "kolu-git";
 import type { TerminalProcess } from "../terminals.ts";
 import { subscribeForTerminal, publishForTerminal } from "../publisher.ts";
 import { updateMetadata } from "./index.ts";
 import { log } from "../log.ts";
 import { trackRecentRepo } from "../state.ts";
-
-const DEBOUNCE_MS = 150;
-
-/** Fast check: does a .git entry exist in this directory? (stat, not a git subprocess) */
-function hasGitDir(cwd: string): boolean {
-  try {
-    fs.accessSync(path.join(cwd, ".git"));
-    return true;
-  } catch {
-    return false;
-  }
-}
-
-/** Resolve git context for a directory. Returns null if not in a git repo. */
-export async function resolveGitInfo(cwd: string): Promise<GitInfo | null> {
-  try {
-    const git = simpleGit(cwd);
-    // Bare repos (core.bare=true) have no work tree, so `--show-toplevel`
-    // throws on them. Detect up front and return a GitInfo rooted at the
-    // bare repo's own location — the palette consumer treats the result as
-    // "a repo you can spawn a worktree from," which is exactly right.
-    const isBare =
-      (await git.raw(["rev-parse", "--is-bare-repository"])).trim() === "true";
-    if (isBare) {
-      // Derive the repo location from `--git-dir`, not cwd. For a canonical
-      // bare repo (`/tmp/foo` bare, cwd == bare dir) the two coincide. For
-      // project layouts where a bare `.git` sits inside a working dir
-      // (`/home/user/proj/.git` with sibling `proj/.worktrees/`), cwd can be
-      // anywhere around `.git` — falling back to `basename(cwd)` would
-      // report the wrong name (e.g. `.worktrees`).
-      const gitDirAbs = fs.realpathSync(
-        path.resolve(cwd, (await git.raw(["rev-parse", "--git-dir"])).trim()),
-      );
-      const gitDirBase = path.basename(gitDirAbs);
-      // Three shapes:
-      //   /proj/.git        → root /proj,        name proj
-      //   /foo.git          → root /foo.git,     name foo
-      //   /foo (bare dir)   → root /foo,         name foo
-      const isDotGit = gitDirBase === ".git";
-      const repoRoot = isDotGit ? path.dirname(gitDirAbs) : gitDirAbs;
-      const repoName = isDotGit
-        ? path.basename(repoRoot)
-        : gitDirBase.replace(/\.git$/, "");
-      let branch: string;
-      try {
-        branch = (await git.raw(["symbolic-ref", "--short", "HEAD"])).trim();
-      } catch {
-        // Detached HEAD in a bare repo (unusual but possible).
-        branch = (await git.revparse(["--abbrev-ref", "HEAD"])).trim();
-      }
-      return {
-        repoRoot,
-        repoName,
-        worktreePath: repoRoot,
-        branch,
-        isWorktree: false,
-        mainRepoRoot: repoRoot,
-      };
-    }
-    const repoRoot = (await git.revparse(["--show-toplevel"])).trim();
-    let branch: string;
-    try {
-      branch = (await git.raw(["symbolic-ref", "--short", "HEAD"])).trim();
-    } catch {
-      branch = (await git.revparse(["--abbrev-ref", "HEAD"])).trim();
-    }
-    // --git-common-dir returns the shared .git dir; for worktrees it points
-    // back to the main repo's .git, letting us derive the real repo name.
-    // The path is relative to cwd (where simple-git runs), not repoRoot.
-    // realpathSync normalizes symlinks (e.g. /tmp → /private/tmp on macOS)
-    // so the comparison with repoRoot (which git already resolved) is reliable.
-    const gitCommonDir = (await git.revparse(["--git-common-dir"])).trim();
-    const mainRepoRoot = path.dirname(
-      fs.realpathSync(path.resolve(cwd, gitCommonDir)),
-    );
-    const isWorktree = mainRepoRoot !== repoRoot;
-    return {
-      repoRoot,
-      repoName: path.basename(mainRepoRoot),
-      worktreePath: cwd,
-      branch,
-      isWorktree,
-      mainRepoRoot,
-    };
-  } catch (err) {
-    // Log so unexpected failures (permission errors, missing git binary)
-    // surface instead of being silently treated as "not a repo".
-    log.debug({ err, cwd }, "git: resolveGitInfo failed");
-    return null;
-  }
-}
-
-/**
- * Watch .git/HEAD for changes (branch switches, checkout, etc.).
- * Returns a cleanup function. Returns a no-op for non-git directories.
- */
-function watchGitHead(cwd: string, onChange: () => void): () => void {
-  let gitDir: string;
-  try {
-    const result = execSync("git rev-parse --git-dir", {
-      cwd,
-      encoding: "utf-8",
-      stdio: ["pipe", "pipe", "pipe"],
-    });
-    gitDir = path.resolve(cwd, result.trim());
-  } catch {
-    return () => {};
-  }
-
-  let timer: ReturnType<typeof setTimeout> | undefined;
-  let watcher: fs.FSWatcher | undefined;
-  try {
-    watcher = fs.watch(gitDir, (_, filename) => {
-      if (filename !== "HEAD") return;
-      if (timer) clearTimeout(timer);
-      timer = setTimeout(onChange, DEBOUNCE_MS);
-    });
-  } catch (err) {
-    log.debug({ err, gitDir }, "git: failed to watch git dir");
-    return () => {};
-  }
-
-  return () => {
-    if (timer) clearTimeout(timer);
-    watcher?.close();
-  };
-}
-
-/** Compare two GitInfo values for equality. */
-export function gitInfoEqual(a: GitInfo | null, b: GitInfo | null): boolean {
-  if (a === b) return true;
-  if (!a || !b) return false;
-  return (
-    a.repoRoot === b.repoRoot &&
-    a.branch === b.branch &&
-    a.worktreePath === b.worktreePath
-  );
-}
 
 /**
  * Start the git metadata provider for a terminal entry.
@@ -171,7 +35,7 @@ export function startGitProvider(
   const plog = log.child({ provider: "git", terminal: terminalId });
   const meta = entry.info.meta;
   let lastCwd = meta.cwd;
-  let stopHeadWatch = watchGitHead(meta.cwd, handleHeadChange);
+  let stopHeadWatch = watchGitHead(meta.cwd, handleHeadChange, plog);
 
   plog.debug({ cwd: lastCwd }, "started");
 
@@ -190,7 +54,7 @@ export function startGitProvider(
     lastCwd = newCwd;
     // Restart HEAD watcher for new directory
     stopHeadWatch();
-    stopHeadWatch = watchGitHead(newCwd, handleHeadChange);
+    stopHeadWatch = watchGitHead(newCwd, handleHeadChange, plog);
     void resolve(newCwd);
   }
 
@@ -200,13 +64,17 @@ export function startGitProvider(
   }
 
   async function resolve(cwd: string) {
-    const git = await resolveGitInfo(cwd);
+    const result = await resolveGitInfo(cwd, plog);
+    const git: GitInfo | null = result.ok ? result.value : null;
+    if (!result.ok && result.error.code !== "NOT_A_REPO") {
+      plog.error({ code: result.error.code }, "git resolution failed");
+    }
     const m = entry.info.meta;
     if (gitInfoEqual(git, m.git)) return;
     // Start HEAD watcher when a repo appears (e.g. after `git init`)
     if (m.git === null && git !== null) {
       stopHeadWatch();
-      stopHeadWatch = watchGitHead(cwd, handleHeadChange);
+      stopHeadWatch = watchGitHead(cwd, handleHeadChange, plog);
     }
     // Track repo in persistent recent repos list
     if (git) trackRecentRepo(git.mainRepoRoot, git.repoName);
